Improved 2dRepresentation

+++ src/Model/BondInferenceAnd2D/Pos2d.java
@@ -6,27 +6,27 @@ package Model.BondInferenceAnd2D;
  */
 public class Pos2d {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Pos2d(int x, int y) {
+    public Pos2d(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
